New release process (again) (#6379)
Follow up to #6365 which was reverted

This time we keep the package.json and version.ts version numbers up to
date as before. I think the cleanest way to do this on the main branch
is to create a new action that can be triggered after each `latest`
release on npm.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`