a11y: add keyboard shortcuts preference (#6363)
From the a11y audit:

> Currently, there are keyboard shortcuts to easily jump to certain
functionality. For example, D activates Darwing, O activates an Oval
shape, N for a new Note, cmd +, Option V aligns vertically, and so on.
> 
> This is useful and allows for quick keyboard navigation (good), but
one of the following should be true for the shortcuts that consist of
only alphanumeric characters (N, O, etc) to include one or more
non-printable keys (like alt, ctrl, shift, etc):
> 
> It should be possible to turn off keyboard shortcuts
> 
> It should be possible to remap keyboard shortcuts that only consist of
> 
> The reason for this is that those are easy to trigger, especially by
voice controls, and might interfere with other assistive technologies
that come with their own shortcuts. People who experience tremors or
other impairments related to coordination might also accidentally press
keys on their keyboard and accidentally trigger a keyboard shortcut by
doing so.
> 
> ### Suggested fixes
> 
> Several options:
> 
> Option A: The quickest fix would be to provide a mechanism (that's
accessible and easily discoverable) to toggle all the keyboard shortcuts
off or on. That way, those who experience issues with them, can just
turn them off. They won't get the benefit of the easy controls, but the
controls won't get in the way of their work either.
> 
> Option B: If keyboard shortcuts are an important feature, then
allowing them to be remapped to include non-printable characters as well
is the more user-sensitive solution. Someone who for example
accidentally keeps hitting the N key while using the trackpad (and by
doing so creating a new note), will be able to still benefit from a
shortcut (quickly creating a new note), but without the accidental
triggering.
> 
> WCAG requires this just for the shortcuts consisting of only
alphanumeric/printable characters. But from a UX point of view, it might
make more sense to allow the remapping/turning off for all shortcuts, to
keep it consistent.
> 
> Also remember: These settings should be remembered when revisiting the
page. A user who needs them to be turned off or remapped, should not
have to go through this flow each time they re-open the page.
> 
> ### Related WCAG
> 
> [SC 2.4.1 Character key shortcuts
(A)](https://www.w3.org/TR/WCAG22/#character-key-shortcuts): If a
keyboard shortcut is implemented in content using only letter,
punctuation, number, or symbol characters, then a mechanism is available
to turn the shortcut off, or a mechanism is available to remap the
shortcut to include one or more non-printable keyboard keys.

Also, there's a recommendation that we add a kbd shortcut _for_ the
keyboard shortcut menu.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [x] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create a shape...
2.

- [x] Unit tests
- [ ] End to end tests

### Release notes

- a11y: add keyboard shortcuts preference