Remove file upsert + legacy mutations (#6400)
The groups pr #6265 will revert the file table primary key to being just
(id). This breaks the `upsert` operation on the file table because it
uses an `on conflict` clause for the whole primary key, and if you
mention non-unique or non-key fields in the `on conflict` clause
postgres throws an error.

Admittedly this error would only be liable to happen for a minute or two
during the deployment, but it affects every file creation flow, so let's
prevent it from being a problem by simply not using `upsert` for the
file creation mutator.

At the same time, I pulled out another change from #6265 to remove the
legacy mutations. since this can be done before that PR and it'll be
easier to review in isolation i think.

### Change type

- [x] `other`

### Test plan

I ran e2e tests on this branch with the groups sql migration applied,
and everything passed.