Fix diagonal arrow issue (#6417)
This fixes #6391. This turned out to be a very sneaky issue where
sometimes, when doing intersections between line segments and polylines,
if the line segment crossed the poly line exactly at one of it's
vertices, floating point errors would cause the intersection not to be
detected.

The fix here is to use our approximate equality helpers to account for
the possibility of floating point errors when working with these
intersections.

### Change type

- [x] `bugfix`

### Test plan

- [x] Unit tests
- [x] End to end tests

### Release notes

- Fix an issue that would cause arrows between circles to sometimes
render incorrectly

---------

Co-authored-by: huppy-bot[bot] <128400622+huppy-bot[bot]@users.noreply.github.com>