Improve frame rendering performance by avoiding re-rendering frame shapes on zoom (#6415)
Instead of reading the zoom level and setting attributes based on it,
use css styles with the --tl-zoom variable instead. This improves
performance a lot when there's many frames on a page.

For example on a page with 256 frames with my laptop in powersave when
I'm zoomed in so just one frame shows and I zoom in/out it goes from
15-20 fps to 30-40 fps. This is just an example, it improves performance
when you're zoomed out and it shows multiple frames too, the fps numbers
are just different in that case.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create lots of frames
2. Zoom in/out

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Improve frame rendering performance by avoiding re-rendering frame
shapes on zoom