Improve select-all action (#6386)
This PR improves the "select all" feature by scoping it to the parent of
the user's current selected shapes, such as all the shapes in a frame
when a user has one of the frame's children selected). If the user has
shapes that belong to multiple parents (e.g. a child of a frame and then
another child of the page) then select all has no effect.

Thanks to @aprocissi for the suggestion. Closes
https://github.com/tldraw/tldraw/issues/6322.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create shapes on the page
3. Select a shape
4. Press cmd+a to Select all
5. The shapes on the page should be selected

1. Create multiple shapes inside of a frame
2. Select one of the shapes in the frame
3. Press cmd+a to Select all
4. All the frame's children should be selected, but nothing else

1. Create a shape inside of a frame
2. Create a shape outside of the frame
3. Select both shapes
4. Press cmd+a to Select all
5. Nothing should be selected

- [x] Unit tests
- [ ] End to end tests

### Release notes

- Improved the behavior of Select All to select within the selected
shape(s) current common parent.