Use slightly lower resolution approximation for PathBuilder arc geometries (#6421)
I noticed when I was working on #6417 that we're using overly high
resolution approximations for arcs since introducing `PathBuilder`.
Before `PathBuilder`, we used an adaptive resolution based on the length
of the arc. After `PathBuilder`, arcs are modelled as 1 to 4 cubic
beziers, each approximated with 10 points.

This diff back-ports the old code we used to calculate how many vertices
to use in an approximation to the new `PathBuilder` stuff. Note the
difference in the number of points used to approximate arcs in these
shapes:

![image](https://github.com/user-attachments/assets/f0477dce-6b30-404f-b300-00af82c036df)


### Change type

- [x] `improvement`

### API changes

- Slightly improve the performance of geometry calculations involving
arcs

---------

Co-authored-by: huppy-bot[bot] <128400622+huppy-bot[bot]@users.noreply.github.com>