Improve navigation panel contrast / icons (#6457)
This pull request includes changes to improve the user interface of the
navigation panel in the `tldraw` application. The most important updates
involve simplifying CSS styling and modifying button components for
better consistency and usability.

### UI Styling Simplifications:
* Removed CSS rules for `.tlui-navigation-panel__toggle .tlui-icon`,
eliminating opacity changes on hover and active states.

### Button Component Updates:
* Updated the `DefaultNavigationPanel` component to use the `small`
variant of `TldrawUiButtonIcon` for the zoom-out button.
* Updated the `DefaultNavigationPanel` component to use the `small`
variant of `TldrawUiButtonIcon` for the zoom-in button.
* Modified the minimap toggle button in `DefaultNavigationPanel` to use
the `small` variant of `TldrawUiButtonIcon` and replaced the icons with
`chevron-right` and `chevron-left` for a more intuitive representation.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Improved icons contrast and design in the navigation panel

---------

Co-authored-by: Mime Čuvalo <mimecuvalo@gmail.com>