switch spinner animations to svg (#6459)
I noticed some weird issues with the SVG animation on our spinner when
working on #6412. The animation wouldn't play immediately, and it looked
like the browser was hanging. Switching to CSS animations fixes this,
and is better anyway because it doesn't force the browser to repaint the
SVG each frame.

I made a couple other related changes while working on this:
1. `@keyframes` declarations in library CSS files are now namespaced
like the class names, so we don't clash with application defined
keyframes
2. the UI `<Spinner />` component now re-uses the editor one

### Change type

- [x] `improvement`

### Test plan

### Release notes

### api changes
- The `<Spinner />` component now accepts SVG props like `className`