allow custom jsx for icons (#6486)
Something that often annoys me when extending our UI is that I can't
pass a custom SVG or whatever when setting icons in our buttons, tools,
and actions etc.

This diff adds support for passing in JSX to use in the place of icons.
There's some trickiness here as our icons sometimes need to be made
small, have aria-labels, or use special classnames, but this can be
worked around with some `cloneElement` juggling.

### Change type

- [x] `improvement`

### Release notes

### API Changes

- You can now pass JSX elements into the `icon` slots of tldraw UI &
overrides