TLSyncRoom optimization (#6488)
We were allocating an awful lot of large collections in this class. This
was a serious problem for many hot paths (updating records being the
main one).

This PR fixes the biggest issues by

- Using AtomMap to store the main document state
- Using lazy iteration when doing many processes that previously used
eagerly-evaluated collections.

While working on this I noticed and fixed an issue with how we handled
tombstone pruning. Basically we weren't updating the
`tombstoneHistoryStartsAtClock` field properly and it left a hole where
deletes might have not been propagated to clients with flaky
connections.

### Change type

- [x] `improvement`

### Release notes

- Improve server-side performance when dealing with large rooms in
tldraw sync

### API changes
- `TLSocketRoom#getRecord` return type now reflects that it returns
`undefined` if the record was not found.
- `StoreSchema#migrateStoreSnapshot` has a new option `mutateInputStore:
boolean` which is false by default. If left false, it will copy the
input store with `structuredClone`. If set to true the mutators will be
allowed to mutate the input store.
- (other internal changes)