add `Editor.updatePointer` method (#6494)
Add `editor.updatePointer`, which dispatched a pointer move event in the
same screen position as the last pointer event the editor received. This
is useful when you want an interaction to update in response to an
external change, such as moving the camera or creating a shape.

### Change type

- [x] `api`

### Release notes

### API Changes
- Add `editor.updatePointer()`, which is useful when you want an
interaction to update in response to an external change, such as moving
the camera or creating a shape.