add custom shape wrapper component (#6514)
Sometimes, it's useful to customize how shapes are rendered without
digging into the shape itself - for example, to apply custom css
classnames to a subset of shapes.

This diff introduces a `ShapeWrapper` component (extracted from the
`Shape` component) and makes it overridable in the same way other editor
components are. It also adds an example showing how this new API can be
used to style a specific shape.

### Change type

- [x] `api`
- [ ] `other`

### Release notes

### API Changes

- Add `ShapeWrapper` to the `components` prop to allow customising how
every single shape is rendered to the DOM.