Improve room history logic. (#6517)
There was an issue with rooms that have a large history. We tried to
fetch all the timestamps (could be thousands) which caused the worker to
take longer than 120s to respond, which means its execution got canceled
due to worker limits.

This is a bit complicated since r2 buckets don't really have a good way
of querying for items. You can't specify in which order you want them
returned and they are always returned from oldest to newest (which is
pretty much the exact opposite of what we would like).

We now:
* Fetch an initial batch of 1000 items from the room history. If we have
fewer than 1000 items we just return them. This keeps the old logic
intact for rooms that don't have a big history.
* For rooms with more than 1000 history entries we start from the
current month and work our way back to find the first month that has
some history entries. Once we find that month we start collecting
history entries. Once we get to at least 1000 (which could span multiple
months), we return those.
* We also check if we have any older entries so that users can request
more.
* We have basic pagination to request more items.

Fixes ENG-3551

### Change type

- [x] `improvement`