add EditorAtom helper (#6531)
When I was working on the workflows starter, I found myself needing to
store a lot of state in atoms. This includes:
* The current exectution
* Which ports on which shapes should be highlighted when dragging a
connection
* Whether the on-canvas component picker was visible, and which shape it
should be attached to

All of this state should really belong to a specific `Editor` instance.
It's referencing data from that instance. If I was building a full app
and navigated from one workflow to another, I would want all the state
to be cleared due to the editor getting re-created. It's possible to get
this sort of functionality by using a weak map from editors to atoms,
but it takes a fair bit of management.

This diff adds an `EditorAtom` helper which handles the weak-map
juggling. The interface is similar to that of `Editor`, but each method
takes an `editor` for scoping.

### Change type

- [x] `api`


### Release notes

### API Changes

- Added `EditorAtom`, a helper for storing editor UI state in an `Atom`