Avoid incorrect session attribution. (#6544)
My understanding is that were were setting a property called `source`
when sending over the events. Since there was no other source set then
GA4 took this as the traffic source attribution as well.

So the problem is not actually using UTM tags for internal navigation.
The problem was that were we using the same name of the property that
GA4 uses for traffic attribution.

### Change type

- [x] `bugfix`