Fix edge hitting on arrows (#6525)
This pull request adds the ability to separate inner and outer margins
for shape detection. It fixes a bug where hollow shapes could not be
bound-to by arrows when overlapping solid filled shapes.


https://github.com/user-attachments/assets/3866a869-90f3-4400-b20d-02774b71cab0

### Change type

- [x] `bugfix`

### Test plan

1. Create a solid shape
2. Create a hollow shape on top
3. Bind an arrow to the hollow shape

- [x] Unit tests

### Release notes

- Fixed a bug where hollow shapes could not be bound-to by arrows when
overlapping a filled shape.

### API changes
- Non-breaking, obscure. The `Editor.getShapesAtPoint` can now accept a
number _or_ a number tuple for its `margin` option that will be used as
the inner and outer margins. This works best when `hitInside` is `true`;
if `hitInside` is false, then the larger of the two margins will be used
for both inside and outside margin.

---------

Co-authored-by: alex <alex@dytry.ch>