Unsaved changes examples (#6533)
This pull request introduces a new example for tracking and handling
unsaved changes in `tldraw` documents. It keeps track of the set of
diffs that have accumulated since the last save point and checks whether
the diff is empty

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [x] `api`

### Release notes

- Add an option to `squashRecordDiffs` to allow mutating the first diff
instead of creating a new diff.

### API changes

- Add an option to `squashRecordDiffs` to allow mutating the first diff
instead of creating a new diff.

---------

Co-authored-by: David Sheldrick <d.j.sheldrick@gmail.com>