Vertical toolbar example (#6532)
Vertical toolbar! Use it like this: 

```tsx
<Tldraw components={{ Toolbar: () => <DefaultToolbar orientation="vertical" /> }} />
```

Right now, I've only supported a vertical toolbar on the left of the
screen because to put it elsewhere we'd need a better way of moving
other parts of the UI around. This also supports groups in the toolbar,
and re-works the overflowing toolbar component to work both horizontally
and vertically.


### Change type

- [x] `api`

### Release notes

### API Changes


- You can now position the toolbar vertically down the right of the
screen by setting it's orientation to `vertical`. You can also now use
`<TldrawUiMenuGroup />` to group together toolbar items.