Add consent form for logged out users. (#6559)
Add cookie consent for logged out users.

1. We store this in local storage.
2. If the user then logs in we sync it with the user's signed in
options, so that they have a consistent experience when logged in or
logged out.
3. If a logged in user changes their preference we also update it in
local storage so that it's the same if they log out.
4. For logged out users we also show an additional entry in the main
menu, which allows them to change their cookie preferences.
5. Decided to go with a more generic storage of the cookie consent,
which can allow us to add consent for cookies with other purposes (like
marketing,...)

We'll adjust the appearance of the consent dialog in a follow up PR.

### Change type

- [x] `improvement`