Add CONTEXT.md files, others (#6588)
This PR adds a bunch of files that should only really matter to LLMs
that are trying to work in our repository.

## Context

This PR adds:
- CONTEXT.md files to all packages
- a `refresh-context` script that will, at great expense, check whether
the CONTEXT.md files are correct

I'd suggest we run this script maybe once a week, since it's just used
as an efficiency / quality improvement for the models. At worst, it'll
be bad directions in the case of large breaking changes or
re-organization within the codebase.

## Other stuff for the bots

Less importantly, it also adds:
- MARKETING_STYLE_GUIDE.md for writing marketing copy / technical copy
- COMPANY.md for information about tldraw itself (why it exists, etc).

Our GH repository is very popular and well-linked, so having information
here is a good idea.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`