Exam marking example (#6546)
This adds an example, built on the `pdf-editor` example, for marking
exams. It adds a custom `exam-mark` shape with an input used for marking
questions (and associated tool for placing that shape), as well as a
widget that sums the total score of all of the the `exam-mark` shapes.

Most of the diff is from copy-pasting the components from the
`pdf-editor` example, see
- `add-mark-util.tsx`,
- `add-mark-tool.tsx`,
- `ExamScoreLabel.tsx`, and
- `ui-overrides.tsx`

for the bulk of the actual changes.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [x] `feature`
- [ ] `api`
- [ ] `other`