Add lasso select tool example (#6564)
Adds an example that shows how to build a lasso select tool. It also
addresses a new api feature: `EditorAtom`s, as well as `Overlays`, (and
the difference between Overlays and `InFrontOfTheCanvas`) which aren't
addressed in any example yet.


![lasso-select-final](https://github.com/user-attachments/assets/84c2463f-d9fd-401d-b331-ba33f05a3a14)

### API Changes
I needed some freehand stuff which hadn't been exported from the
`tldraw` package yet, @SomeHats seemed to think this was an oversight so
I exported it and included that in this PR. If that's an issue then lmk
and I will fix :)

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [x] `feature`
- [x] `api`
- [ ] `other`