Export `doesGeometryOverlapPolygon` (#6679)
This PR exports `doesGeometryOverlapPolygon` publicly. The agent starter
makes use of it.

In the end, this is the only change to the SDK needed by the agent
starter as we removed the dependency on the AI module. We'll still need
to fix canvas pointer events when multiple editors are open on the same
page, but that's an SDK fix, not a feature.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [x] `api`
- [ ] `other`

### Release notes

- Added the `overlapsPolygon` method to the `Geometry2D` class.

### API Changes

- Added the `overlapsPolygon` method to the `Geometry2D` class.