More customizable style panel (#6672)
The code for the style panel was more complex than it needed to be.
There was a bunch of rendering logic applied with JS that's doable with
CSS, and a bunch of repetition and prop drilling that could be replaced
with a single context object. The result is a style panel that you can
customize much more easily, following the pattern of our toolbar. To
customize the style panel in your app, you would now paste this snippet:
```tsx
function CustomStylePanelContent() {
	const { styles } = useStylePanelContext()
	if (!styles) return null

	return (
		<>
			<StylePanelSection>
				<StylePanelColorPicker />
				<StylePanelOpacityPicker />
			</StylePanelSection>
			<StylePanelSection>
				<StylePanelFillPicker />
				<StylePanelDashPicker />
				<StylePanelSizePicker />
			</StylePanelSection>
			<StylePanelSection>
				<StylePanelFontPicker />
				<StylePanelTextAlignPicker />
				<StylePanelLabelAlignPicker />
			</StylePanelSection>
			<StylePanelSection>
				<StylePanelGeoShapePicker />
				<StylePanelArrowKindPicker />
				<StylePanelArrowheadPicker />
				<StylePanelSplinePicker />
			</StylePanelSection>
		</>
	)
}

const components = {
	StylePanel: () => <DefaultStylePanel><CustomStylePanelContent /></DefaultStylePanel>,
}
```

and start editing that code to suit your needs. I still want to add a
proper example for this.

### Change type

- [x] `api`

### Release notes

### API Changes

The style panel can now be deeply customised in the same way as the
toolbar component. **BREAKING:** many old components from the style
panel have been removed in favour of newly reorganised ones.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>