chat template: add uploads, temporarily disable image output (#6715)
Vercel has a limit of 4.5mb on requests. the images that gemini creates
are pretty large - we reach that _very_ quickly. to work around this, we
now upload images to gemini's file hosting service instead of sending
them with every request. for some reason though, we're not allowed to do
that with messages from the assistant - they have to be `data:` urls. so
for now, we need to disable image output :(

### Change type

- [x] `other`