shape ordering: use jittered-fractional-indexing package (#6646)
We added jittered indexes in this PR
https://github.com/tldraw/tldraw/pull/4312 building off of strict types
in this PR https://github.com/tldraw/tldraw/pull/2827

This definitely helped with collisions that we were having, esp. pages
with a large # of shapes.

However, the package we were using
(`[fractional-indexing-jittered](https://github.com/TMeerhof/fractional-indexing-jittered/tree/main)`)
wasn't perfect. As described astutely in one ticket, "it doesn't inspire
confidence".

@nathanhleung took it upon himself to create a new fork of the original
package and created:
https://github.com/nathanhleung/jittered-fractional-indexing
which is based off of
https://github.com/rocicorp/fractional-indexing

which after review does inspire more confidence.

Here are the related issues:
- https://github.com/tldraw/tldraw/issues/5864
- https://github.com/tldraw/tldraw/issues/6141



### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

- [x] Unit tests
- [ ] End to end tests

### Release notes

- shape ordering: use jittered-fractional-indexing package