Add maximum file size in description of the file too big toast (#6745)
When a user tries to upload a file that is larger than `maxAssetSize`,
it would be useful to add a description of what that size is. This PR
adds that description to the toast that is triggered by this behaviour.

## Before
<img width="232" height="64" alt="image"
src="https://github.com/user-attachments/assets/fb3c7f43-beb8-416d-a174-b160583c3fd8"
/>

## After
<img width="276" height="83" alt="image"
src="https://github.com/user-attachments/assets/78fdf560-da89-4470-b543-9cab4085317e"
/>

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### API changes
- Added a new translation key `assets.files.maximum-size`

### Release notes

- Added maximum file size to the description of the file too big toast,
when uploading a file.